pyproject.toml
[tool.poetry]
name = "ip2loc"
version = "0.1.0"
description = ""
authors = ["kuchichan <pawel.kucharski@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
django-environ = "^0.4.5"
djangorestframework = "^3.12.2"
requests = "^2.25.1"
djangorestframework-simplejwt = "^4.6.0"
pytest-mock = "^3.5.1"
django-stubs = "^1.7.0"
ipython = "^7.21.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest-django = "^4.1.0"
factory-boy = "^3.2.0"
ipython = "^7.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
